feat(crud): gestion de l'ajout, de la modification et de la suppression

+++ src/Controller/Admin/UserController.php
@@ -89,14 +89,20 @@ class UserController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_admin_user_delete', methods: ['POST'])]
+    #[Route('/{id}/delete', name: 'app_admin_user_delete', methods: [Request::METHOD_GET, Request::METHOD_POST])]
     public function delete(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
-            $entityManager->remove($user);
-            $entityManager->flush();
+        if ($request->getMethod() == Request::METHOD_POST) {
+            if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
+                $entityManager->remove($user);
+                $entityManager->flush();
+            }
+
+            return $this->redirectToRoute('app_admin_user_index', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->redirectToRoute('app_admin_user_index', [], Response::HTTP_SEE_OTHER);
+        return $this->renderForm('admin/user/delete.html.twig', [
+            'user' => $user,
+        ]);
     }
 }
